# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "DocumentResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "failure_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Reason"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "latest_revision_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Revision Id"
          },
          "matter_id": {
            "title": "Matter Id",
            "type": "string"
          },
          "provenance_state": {
            "title": "Provenance State",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "matter_id",
          "title",
          "status",
          "provenance_state",
          "created_at"
        ],
        "title": "DocumentResponse",
        "type": "object"
      },
      "DownloadResponse": {
        "properties": {
          "expires_in_seconds": {
            "title": "Expires In Seconds",
            "type": "integer"
          },
          "presigned_url": {
            "title": "Presigned Url",
            "type": "string"
          }
        },
        "required": [
          "presigned_url",
          "expires_in_seconds"
        ],
        "title": "DownloadResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MatterResponse": {
        "properties": {
          "access_scope": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Access Scope"
          },
          "ai_processing_policy": {
            "title": "Ai Processing Policy",
            "type": "string"
          },
          "case_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Case Type"
          },
          "client_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Name"
          },
          "closed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closed At"
          },
          "confidentiality_level": {
            "title": "Confidentiality Level",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "internal_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Internal Reference"
          },
          "jurisdiction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Jurisdiction"
          },
          "opened_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opened At"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "status",
          "confidentiality_level",
          "ai_processing_policy"
        ],
        "title": "MatterResponse",
        "type": "object"
      },
...
      "ParsedDocumentResponse": {
        "properties": {
          "document_id": {
            "title": "Document Id",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "parser_used": {
            "title": "Parser Used",
            "type": "string"
          },
          "parsing_revision": {
            "title": "Parsing Revision",
            "type": "integer"
          },
          "revision_id": {
            "title": "Revision Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "document_id",
          "revision_id",
          "parsing_revision",
          "parser_used",
          "status"
        ],
        "title": "ParsedDocumentResponse",
        "type": "object"
      },
...
      "UploadCompleteResponse": {
        "properties": {
          "revision_id": {
            "title": "Revision Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "revision_id"
        ],
        "title": "UploadCompleteResponse",
        "type": "object"
      },
      "UploadIntentRequest": {
        "properties": {
          "filename": {
            "maxLength": 255,
            "title": "Filename",
            "type": "string"
          },
          "matter_id": {
            "title": "Matter Id",
            "type": "string"
          },
          "mime_type": {
            "maxLength": 100,
            "title": "Mime Type",
            "type": "string"
          },
          "size_bytes": {
            "description": "File size in bytes",
            "exclusiveMinimum": 0.0,
            "maximum": 104857600.0,
            "title": "Size Bytes",
            "type": "integer"
          }
        },
        "required": [
          "matter_id",
          "filename",
          "mime_type",
          "size_bytes"
        ],
        "title": "UploadIntentRequest",
        "type": "object"
      },
      "UploadIntentResponse": {
        "properties": {
          "document_id": {
            "title": "Document Id",
            "type": "string"
          },
          "presigned_url": {
            "title": "Presigned Url",
            "type": "string"
          },
          "revision_id": {
            "title": "Revision Id",
            "type": "string"
          },
          "storage_key": {
            "title": "Storage Key",
            "type": "string"
          }
        },
        "required": [
          "document_id",
          "revision_id",
          "presigned_url",
          "storage_key"
        ],
        "title": "UploadIntentResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "MESA Law API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/documents/matters/{matter_id}": {
      "get": {
        "operationId": "listMatterDocuments",
        "parameters": [
          {
            "in": "path",
            "name": "matter_id",
            "required": true,
            "schema": {
              "title": "Matter Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DocumentResponse"
                  },
                  "title": "Response Listmatterdocuments",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Matter Documents"
      }
    },
    "/api/v1/documents/upload-intent": {
      "post": {
        "operationId": "createUploadIntent",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadIntentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadIntentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Upload Intent"
      }
    },
    "/api/v1/documents/{document_id}": {
      "get": {
        "operationId": "getDocument",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Document"
      }
    },
    "/api/v1/documents/{document_id}/complete": {
      "post": {
        "operationId": "completeUpload",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadCompleteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Complete Upload"
      }
    },
    "/api/v1/documents/{document_id}/download": {
      "get": {
        "operationId": "downloadDocument",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DownloadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Document"
      }
    },
...
    "/api/v1/matters/{matter_id}": {
      "get": {
        "operationId": "getMatter",
        "parameters": [
          {
            "in": "path",
            "name": "matter_id",
            "required": true,
            "schema": {
              "title": "Matter Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatterResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Matter"
      }
    },
...
    "/api/v1/parser/document/{document_id}": {
      "get": {
        "operationId": "listParsedDocuments",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ParsedDocumentResponse"
                  },
                  "title": "Response Listparseddocuments",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Parsed Documents"
      }
    },
...
  }
}